Improve boot reliability on qemux86.
:Release Notes:
In qemux86, novacomd doesn't always start correctly as the ip address
has not yet been set. Current code adds a fixed 3 second delay.

:Detailed Notes:
Revise the Upstart job for emulator to try to start novacomd up to 10
times with a one-second sleep between attempts.  If the last pass
still doesn't work, dump an error to the console and exit.
Another problem showed up where there is a second route shown with
"UGH" in the flags and a "$2" of an IP address.  This was getting
joined on to the end of the command line and confused the shell
which caused the job to fail. Changing the grep to look for "UG "
fixed this issue.

:Testing Performed:
Verification build tested with overnight automatic looping job.

:QA Notes:

:Issues Addressed:
[CTO-640] Fine tune boot sequence

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: Id9f999b50a012c6d0c99cb3cc27406632a87cd25
Reviewed-on: https://g2g.lgsvl.com/6465
Reviewed-by: DCO Verification
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Steve Lemke <steve.lemke@lge.com>